size_t to unsigned int

+++ larpandoracontent/LArHelpers/LArDiscreteProbabilityHelper.cc
@@ -13,14 +13,14 @@ namespace lar_content
 
 template <typename T>
 float LArDiscreteProbabilityHelper::CalculateCorrelationCoefficientPValueFromPermutationTest(const T &t1, const T &t2, 
-    std::mt19937 &randomNumberGenerator, const size_t nPermutations)
+    std::mt19937 &randomNumberGenerator, const unsigned int nPermutations)
 {
     if (1 > nPermutations)
         throw pandora::StatusCodeException(pandora::STATUS_CODE_INVALID_PARAMETER);
 
     const float rNominal(LArDiscreteProbabilityHelper::CalculateCorrelationCoefficient(t1,t2));
 
-    int nExtreme(0);
+    unsigned int nExtreme(0);
     for (unsigned int iPermutation = 0; iPermutation < nPermutations; ++iPermutation)
     {
         const float rRandomised(LArDiscreteProbabilityHelper::CalculateCorrelationCoefficient(
@@ -38,7 +38,7 @@ float LArDiscreteProbabilityHelper::CalculateCorrelationCoefficientPValueFromPer
 
 template <typename T>
 float LArDiscreteProbabilityHelper::CalculateCorrelationCoefficientPValueFromStudentTDistribution(const T &t1, 
-    const T &t2, const size_t nIntegrationSteps)
+    const T &t2, const unsigned int nIntegrationSteps)
 {
     const float correlation(LArDiscreteProbabilityHelper::CalculateCorrelationCoefficient(t1,t2));
     const float dof(static_cast<float>(LArDiscreteProbabilityHelper::GetSize(t1)) - 2.f);
@@ -73,7 +73,7 @@ float LArDiscreteProbabilityHelper::CalculateCorrelationCoefficient(const T &t1,
 
     float variance1(0.f), variance2(0.f), covariance(0.f);
 
-    for (size_t iElement = 0; iElement < LArDiscreteProbabilityHelper::GetSize(t1); ++iElement)
+    for (unsigned int iElement = 0; iElement < LArDiscreteProbabilityHelper::GetSize(t1); ++iElement)
     {
         const float diff1(LArDiscreteProbabilityHelper::GetElement(t1,iElement) - mean1);
         const float diff2(LArDiscreteProbabilityHelper::GetElement(t2,iElement) - mean2);
@@ -99,7 +99,7 @@ float LArDiscreteProbabilityHelper::CalculateMean(const T &t)
         throw pandora::StatusCodeException(pandora::STATUS_CODE_NOT_INITIALIZED);
 
     float mean(0.f);
-    for (size_t iElement = 0; iElement < LArDiscreteProbabilityHelper::GetSize(t); ++iElement)
+    for (unsigned int iElement = 0; iElement < LArDiscreteProbabilityHelper::GetSize(t); ++iElement)
         mean+=LArDiscreteProbabilityHelper::GetElement(t,iElement);
 
     mean /= static_cast<float>(LArDiscreteProbabilityHelper::GetSize(t));
@@ -110,14 +110,14 @@ float LArDiscreteProbabilityHelper::CalculateMean(const T &t)
 //------------------------------------------------------------------------------------------------------------------------------------------
 
 template float LArDiscreteProbabilityHelper::CalculateCorrelationCoefficientPValueFromPermutationTest(const DiscreteProbabilityVector &, 
-    const DiscreteProbabilityVector &, std::mt19937 &, const size_t);
+    const DiscreteProbabilityVector &, std::mt19937 &, const unsigned int);
 template float LArDiscreteProbabilityHelper::CalculateCorrelationCoefficientPValueFromPermutationTest(const std::vector<float> &, 
-    const std::vector<float> &, std::mt19937 &, const size_t);
+    const std::vector<float> &, std::mt19937 &, const unsigned int);
 
 template float LArDiscreteProbabilityHelper::CalculateCorrelationCoefficientPValueFromStudentTDistribution(
-    const DiscreteProbabilityVector &, const DiscreteProbabilityVector &, const size_t);
+    const DiscreteProbabilityVector &, const DiscreteProbabilityVector &, const unsigned int);
 template float LArDiscreteProbabilityHelper::CalculateCorrelationCoefficientPValueFromStudentTDistribution(const std::vector<float> &, 
-    const std::vector<float> &, const size_t);
+    const std::vector<float> &, const unsigned int);
 
 template float LArDiscreteProbabilityHelper::CalculateCorrelationCoefficient(const DiscreteProbabilityVector &, 
     const DiscreteProbabilityVector &);
